Update documentation to reflect current build information
The previous commit (5b0ef877091b7662e7ee26ab8f66058b6ef94977 "Add
support for 390.x driver and refactor existing code") did not update
README.md to mention the consolidated libdir variable nor the new
default values for libdir and TARGET_VER, so this brings it in sync with
those changes.

This also mentions some common default values for libdir and makes note
of support for 331.x which was previously only referenced in nvml_fix.c
and the commit log.

For Debian-based distros using a diversion is a better alternative to
simply deleting the symlink, so the README now includes an example of
how to use dpkg-divert.

Some formatting and readability changes were made as well.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>