Add support for 410.x driver series
This adds offsets for the 410.x series, which is superseding 390.x as
NVIDIA's new long-lived branch.

The default TARGET_VER has been changed to 410.73, the latest 410.x
driver currently available in Ubuntu's "graphics-drivers" PPA.

As is the case with the last few supported major versions, only
offsets for x86_64 have been added.

Tested on Ubuntu 18.10 (amd64) with a GeForce GTX 650 Ti.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>