Add new nvml.h file
This was originally being pulled from the system include dir for
versions >= 390.x, but not all distros make this header easily
available.

The file has been renamed to nvml_v9.h to represent the NVML_API_VERSION
in the file.

The new header was taken from nvidia-cuda-toolkit 9.1.85-4ubuntu1 on
Ubuntu 18.10 amd64.

This also renames the existing nvml.h to nvml_v3.h for consistency.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>